+++ KmpJillcess/src/appleMain/kotlin/com/oldguy/jillcess/cryptography/OfficeAgile.kt
@@ -1,6 +1,6 @@
-package com.oldguy.jillcess.cryptography
+package com.oldguy.jillcess
 
-import com.oldguy.jillcess.OfficeAgileAppleParser
+import com.oldguy.jillcess.cryptography.CTEncryption
 
 actual class OfficeAgile actual constructor() {
     actual fun parseXml(xml: String): CTEncryption {