+++ KmpJillcess/src/commonMain/kotlin/com/oldguy/jillcess/implementations/FileStructureProperties.kt
@@ -1,8 +1,7 @@
 package com.oldguy.jillcess.implementations
 
-import com.oldguy.common.io.ByteBuffer
-import com.oldguy.common.io.Charset
 import com.oldguy.common.io.UByteBuffer
+import com.oldguy.common.io.charsets.Charset
 import com.oldguy.jillcess.toHexString
 
 class LvProperties(