+++ KmpJillcess/src/commonMain/kotlin/com/oldguy/xml/XmlNode.kt
@@ -1,0 +1,32 @@
+package com.oldguy.xml
+
+data class Attribute(
+    val name: String,
+    val value: String
+)
+
+class Node(
+    val name: String,
+    val value : String,
+    val attributes: List<Attribute>,
+    val children: List<Node>
+) {
+    fun find(name: String): Node? {
+        if (this.name == name)
+            return this
+        children.forEach {
+            val result = it.find(name)
+            if (result != null)
+                return result
+        }
+        return null
+    }
+
+    fun findAttribute(name: String): Attribute? {
+        attributes.forEach {
+            if (it.name == name)
+                return it
+        }
+        return null
+    }
+}